+++ frontend/cypress/package.json
@@ -5,6 +5,6 @@
     "debug": "cypress open"
   },
   "devDependencies": {
-    "cypress": "^13.3.0"
+    "cypress": "^13.6.4"
   }
 }
