+++ frontend/package.json
@@ -15,54 +15,56 @@
   "main": "index.js",
   "dependencies": {
     "@fontsource/titillium-web": "^4.5.3",
-    "@fortawesome/fontawesome-svg-core": "^6.0.0",
-    "@fortawesome/free-brands-svg-icons": "^6.0.0",
-    "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@fortawesome/vue-fontawesome": "^3.0.3",
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/free-brands-svg-icons": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@fortawesome/vue-fontawesome": "^3.0.6",
     "@types/markdown-it": "^13.0.7",
     "@types/minimatch": "^5.1.2",
-    "@types/seedrandom": "^3.0.5",
-    "core-js": "^3.6.5",
+    "@types/seedrandom": "^3.0.8",
+    "core-js": "^3.36.1",
     "highlight.js": "^10.5.0",
     "jszip": "^3.5.0",
-    "markdown-it": "^14.0.0",
+    "markdown-it": "^14.1.0",
     "minimatch": "^5.0.1",
     "muicss": "^0.10.3",
     "normalize.css": "^8.0.1",
+    "postcss": "^8.4.38",
     "pug-lint-vue": "^0.4.0",
     "seedrandom": "^3.0.5",
-    "vue": "^3.2.47",
-    "vue-loading-overlay": "^5.0.3",
+    "vue": "~3.4.0",
+    "vue-loader": "^17.4.2",
+    "vue-loading-overlay": "^6.0.4",
     "vue-observe-visibility": "^1.0.0",
-    "vue-router": "^4.1.6",
+    "vue-router": "^4.3.0",
     "vuex": "^4.0.2",
-    "zod": "^3.22.3"
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.17.0",
+    "@babel/eslint-parser": "^7.24.1",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "@vue/cli-plugin-babel": "^5.0.4",
     "@vue/cli-plugin-eslint": "^5.0.4",
     "@vue/cli-plugin-typescript": "^5.0.4",
-    "@vue/cli-service": "^5.0.4",
-    "@vue/compiler-sfc": "^3.2.47",
+    "@vue/cli-service": "^5.0.8",
+    "@vue/compiler-sfc": "^3.4.21",
     "@vue/eslint-config-typescript": "^10.0.0",
-    "eslint": "^8.33.0",
+    "eslint": "^8.57.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-plugin-import": "^2.26.0",
+    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-vue": "^8.4.1",
-    "postcss-html": "^1.4.1",
+    "postcss-html": "^1.6.0",
     "pug": "^3.0.2",
     "pug-plain-loader": "^1.1.0",
-    "sass": "^1.58.0",
-    "sass-loader": "^12.6.0",
+    "sass": "^1.71.0",
+    "sass-loader": "^13.3.3",
     "stylelint": "^14.7.1",
     "stylelint-config-recommended-scss": "^6.0.0",
-    "stylelint-config-recommended-vue": "^1.4.0",
+    "stylelint-config-recommended-vue": "^1.5.0",
     "stylelint-order": "^5.0.0",
-    "stylelint-scss": "^4.2.0",
+    "stylelint-scss": "^4.7.0",
     "stylelint-webpack-plugin": "^3.2.0",
     "typescript": "^4.9.5"
   },
