+++ frontend/src/app.vue
@@ -87,7 +87,7 @@ const app = defineComponent({
           window.REPORT_ZIP = zip;
         }, () => {
           window.alert('Either the .zip file is corrupted, or you uploaded a .zip file that is not generated '
-              + 'by RepoSense.');
+            + 'by RepoSense.');
         })
         .then(() => this.updateReportView());
     },
@@ -113,6 +113,7 @@ const app = defineComponent({
           reportGenerationTime,
           errorMessages,
           names,
+          blurbMap,
         } = summary;
         this.creationDate = creationDate;
         this.reportGenerationTime = reportGenerationTime;
@@ -125,6 +126,7 @@ const app = defineComponent({
         this.getUsers();
         this.renderTabHash();
         this.userUpdated = true;
+        this.$store.commit('setBlurbMap', blurbMap);
       } catch (error) {
         window.alert(error);
       } finally {
