+++ frontend/src/components/c-title.vue
@@ -1,12 +1,16 @@
 <template lang="pug">
-.title(v-html="markdownText", v-if="markdownText != ''")
+.title
+  c-markdown-chunk(v-bind:markdown-text="markdownText")
 </template>
 
 <script lang="ts">
-import MarkdownIt from 'markdown-it';
 import { defineComponent } from 'vue';
+import cMarkdownChunk from './c-markdown-chunk.vue';
 
 export default defineComponent({
+  components: {
+    cMarkdownChunk,
+  },
   data() {
     return {
       markdownText: '',
@@ -19,8 +23,7 @@ export default defineComponent({
       }
       return response.text();
     }).then((text) => {
-      const md = new MarkdownIt({ html: true });
-      this.markdownText = md.render(text);
+      this.markdownText = text;
     }).catch((error) => {
       this.markdownText = (error as Error).toString();
     });
@@ -30,8 +33,6 @@ export default defineComponent({
 
 <style lang="scss" scoped>
 .title {
-  overflow-x: auto;
-  padding: 0 1.5rem;
   // This is needed because the parent summary-wrapper center aligns everything
   text-align: initial;
 }
